Buzzflix/movies/serializers.py
from rest_framework import serializers

from .models import Category, Movie

class MovieSerializer(serializers.ModelSerializer):
    class Meta:
        model = Movie
        fields = ('id', 'title', 'poster_image')

class CategorySerializer(serializers.ModelSerializer):
    movies = MovieSerializer(many=True)

    class Meta:
        model = Category
        fields = ('name', 'movies')