requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-debug-toolbar==3.2
python-dateutil==2.8.1